+++ src/sc_crawler/lookup.py
@@ -1,3 +1,5 @@
+from typing import List
+
 from .tables import ComplianceFramework, Country
 
 # country codes: https://en.wikipedia.org/wiki/ISO_3166-1#Codes
@@ -54,9 +56,15 @@ compliance_frameworks: dict = {
         description="SOC 2 Type 2 is a framework for assessing and certifying the effectiveness of a service organization's information security policies and procedures over time, emphasizing the operational aspects and ongoing monitoring of controls.",  # noqa: E501
         homepage="https://www.aicpa-cima.com/topic/audit-assurance/audit-and-assurance-greater-than-soc-2",  # noqa: E501
     ),
+    "iso27001": ComplianceFramework(
+        compliance_framework_id="iso27001",
+        name="ISO/IEC 27001",
+        abbreviation="ISO 27001",
+        description="ISO 27001 is standard for information security management systems.",  # noqa: E501
+        homepage="https://www.iso.org/standard/27001",  # noqa: E501
+    ),
     # TODO add more e.g.
     # soc2t1
-    # iso27001
     # iso27701
     # gdpr
     # pci
@@ -64,3 +72,26 @@ compliance_frameworks: dict = {
     # csa
 }
 """Dictionary of [sc_crawler.tables.ComplianceFramework][] instances keyed by the `compliance_framework_id`."""
+
+
+def map_compliance_frameworks_to_vendor(
+    vendor_id: str, compliance_framework_ids: List[str]
+) -> dict:
+    """Map compliance frameworks to vendors in a dict.
+
+    Args:
+        vendor_id: identifier of a [Vendor][sc_crawler.tables.Vendor]
+        compliance_framework_ids: identifier(s) of [`ComplianceFramework`][sc_crawler.tables.ComplianceFramework]
+
+    Returns:
+        Array of dictionaroes that can be passed to [sc_crawler.insert.insert_items][].
+    """
+    items = []
+    for compliance_framework_id in compliance_framework_ids:
+        items.append(
+            {
+                "vendor_id": vendor_id,
+                "compliance_framework_id": compliance_framework_id,
+            }
+        )
+    return items